05_Basic_LG_Agent/pyproject.toml
[project]
name = "basic-lg-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langgraph",
    "python-dotenv",
    "pydantic",
    "langchain_google_genai",
    "langchain"
]